build: rename oopsy back to oopsy_live
